Add serializable dictionary for npc locations by date and update position logic
