+++ Assets/Scripts/Data/DataModels/NPCs/NPCLocationsByDate.cs
@@ -1,0 +1,6 @@
+﻿using System;
+
+[Serializable]
+public class NPCLocationsByDate : SerializableDictionary<Date, NPCLocation>
+{
+}
