+++ Assets/Scripts/Data/DataModels/NPCs/NPCLocationsByDate.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 56cdd501ff991bc4d8c86a1224ad7314
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
