+++ Assets/Scripts/Data/SerializableDictionary.cs
@@ -1,0 +1,48 @@
+﻿using System;
+using System.Collections.Generic;
+using UnityEngine;
+
+/// <summary>
+/// Allows a List of keys and a List of values to appear in the inspector as a Dictionary. 
+/// </summary>
+/// <typeparam name="TKey"></typeparam>
+/// <typeparam name="TValue"></typeparam>
+[Serializable]
+public class SerializableDictionary<TKey, TValue> : Dictionary<TKey, TValue>, ISerializationCallbackReceiver
+{
+    [SerializeField]
+    private List<TKey> keys = new List<TKey>();
+
+    [SerializeField]
+    private List<TValue> values = new List<TValue>();
+
+    // Save the Dictionary to Lists
+    public void OnBeforeSerialize()
+    {
+        keys.Clear();
+        values.Clear();
+
+        foreach (KeyValuePair<TKey, TValue> kvp in this)
+        {
+            keys.Add(kvp.Key);
+            values.Add(kvp.Value);
+        }
+    }
+
+    // Load Dictionary from Lists
+    public void OnAfterDeserialize()
+    {
+        Clear();
+
+        if (keys.Count != values.Count)
+        {
+            Debug.LogError($"Dictionary key count ({keys.Count}) does not match values ({values.Count})");
+            return;
+        }
+
+        for (int i = 0; i < keys.Count; i++)
+        {
+            Add(keys[i], values[i]);
+        }
+    }
+}
